requirements_dev.txt
black==23.3.0
faiss-cpu==1.10.0
jupyterlab==4.2.5
langchain-community==0.3.20
langchain-groq==0.2.0
langchain-huggingface==0.1.0
langchain-openai==0.2.3
langchainhub==0.1.21
matplotlib==3.7.1
numpy==1.26.4
pandas==2.0.1
parsenvy==3.0.2
protobuf==4.23.3
pypdf==4.2.0
python-dotenv==1.0.1
seaborn==0.11.2
scikit-learn==1.2.2
statsmodels==0.14.0
pytest==7.4.0
testbook==0.4.2
gradio






